Update README.md
Just an edit for punctuation. companys -> companies.
I tend to do this a lot in my projects as well..... 

It would be wise to include a message saying that Wyze Light Switches are not supported as of now. Why I came here... I saw a reddit post saying that soon it will be.